+++ CHANGELOG.md
@@ -3,7 +3,9 @@
 ## [v0.5.0](https://github.com/agrande/lba2remake/compare/v0.4.0...v0.5.0) Big update!
 _Unreleased_
 
-TODO: Fill this
+#### Editor
+
+- Palette: Eye drop colors from any editor's main view.
 
 ## [v0.4.0](https://github.com/agrande/lba2remake/compare/v0.3.0...v0.4.0) Big update!
 _October 10th, 2020_